refactoring based on comments
