+++ validator/node/node_test.go
@@ -5,6 +5,8 @@ import (
 	"flag"
 	"fmt"
 	"io/ioutil"
+	"net/http"
+	"net/http/httptest"
 	"os"
 	"path/filepath"
 	"testing"
@@ -183,4 +185,126 @@ func newWeb3SignerCli(t *testing.T, baseUrl string, publicKeysOrURL string) *cli
 	require.NoError(t, set.Set(flags.Web3SignerURLFlag.Name, baseUrl))
 	require.NoError(t, set.Set(flags.Web3SignerPublicValidatorKeysFlag.Name, publicKeysOrURL))
 	return cli.NewContext(&app, set, nil)
+}
+
+type test struct {
+	Foo string `json:"foo"`
+	Bar int    `json:"bar"`
+}
+
+func TestUnmarshalFromFile(t *testing.T) {
+	ctx := context.Background()
+	type args struct {
+		File string
+		To   interface{}
+	}
+	tests := []struct {
+		name        string
+		args        args
+		want        interface{}
+		urlResponse string
+		wantErr     bool
+	}{
+		{
+			name: "Happy Path File",
+			args: args{
+				File: "./testdata/test-unmarshal-good.json",
+				To:   &test{},
+			},
+			want: &test{
+				Foo: "foo",
+				Bar: 1,
+			},
+			wantErr: false,
+		},
+		{
+			name: "Bad File Path, not json",
+			args: args{
+				File: "./jsontools.go",
+				To:   &test{},
+			},
+			want:    &test{},
+			wantErr: true,
+		},
+		{
+			name: "Bad File Path",
+			args: args{
+				File: "./testdata/test-unmarshal-bad.json",
+				To:   &test{},
+			},
+			want:    &test{},
+			wantErr: true,
+		},
+		{
+			name: "Bad File Path, not found",
+			args: args{
+				File: "./test-notfound.json",
+				To:   &test{},
+			},
+			want:    &test{},
+			wantErr: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if err := unmarshalFromFile(ctx, tt.args.File, tt.args.To); (err != nil) != tt.wantErr {
+				t.Errorf(" error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			require.DeepEqual(t, tt.want, tt.args.To)
+		})
+	}
+}
+
+func TestUnmarshalFromURL(t *testing.T) {
+	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(200)
+		w.Header().Set("Content-Type", "application/json")
+		_, err := fmt.Fprintf(w, `{ "foo": "foo", "bar": 1}`)
+		require.NoError(t, err)
+	}))
+	defer srv.Close()
+	ctx := context.Background()
+	type args struct {
+		URL string
+		To  interface{}
+	}
+	tests := []struct {
+		name        string
+		args        args
+		want        interface{}
+		urlResponse string
+		wantErr     bool
+	}{
+		{
+			name: "Happy Path URL",
+			args: args{
+				URL: srv.URL,
+				To:  &test{},
+			},
+			want: &test{
+				Foo: "foo",
+				Bar: 1,
+			},
+			wantErr: false,
+		},
+		{
+			name: "Bad URL",
+			args: args{
+				URL: "sadjflksdjflksadjflkdj",
+				To:  &test{},
+			},
+			want:    &test{},
+			wantErr: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if err := unmarshalFromURL(ctx, tt.args.URL, tt.args.To); (err != nil) != tt.wantErr {
+				t.Errorf(" error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			require.DeepEqual(t, tt.want, tt.args.To)
+		})
+	}
 }
